Set NS_port on UNIX client sockets
This patch populates the NS_port property on client Proto::UNIX sockets,
by copying the value from the server socket on accept, which is the same behavior as
the INET sockets.

amavisd-new depends on the NS_port value to distinguish which socket has accepted
a connection to apply a per-socket configuration such as the network ACL and the protocol.

ref:
- https://rt.cpan.org/Public/Bug/Display.html?id=117422
- https://lists.amavis.org/pipermail/amavis-users/2015-April/003549.html