Add SSL_verify_callback to @ssl_args
In a situation where we want to implement TOFU from the server side, accepting self-signed client certificates and storing their fingerprints, we need to tell the server to verify the peer certificates (SSL_verify_mode => SSL_VERIFY_PEER) and we need to write the verification code ourselves (SSL_verify_callback => \&verify_client_cert). Without this change, SSL_verify_mode is passed on to IO::Socket::SSL but SSL_verify_callback is not, resulting in the server verifying and rejecting our self-signed client certificates.