Fix spin loop in SSLEAY::print
If a client becomes totally unresponsive, SSLEAY will eventually transmit a 500/server error regarding the timeout reading headers. For an unresponsive client, this will eventually time out, and the server process will spin with select returning EAGAIN/EWOULDBLOCK.

The fix is to return from 'print' if the $client filehandle is at eof.