fix UDP recieving in Fork server
The return value of getsockopt should be used directly, not attempting
to convert with unpack. This matches how the other server classes do.

This bug was introduced in 5a689da0852d550ab301f0994b8a7820bba157a9
which was otherwise unrelated. The intent of the change in not obvious.

Fixes RT#146575